Calculates N-Sink nitrogen removal percent

Starting with base data layers of NHDPlus, SSURGO, impervious surface, flow velocity, and time of travel, this function calculates percentage of Nitrogen removal. Nitrogen removal methods and calculations are from Kellogg et al. (2010). This function assumes data has been downloaded with nsink_get_data and has been prepared with nsink_prep_data.

Usage

nsink_calc_removal(
  input_data,
  off_network_lakes = NULL,
  off_network_streams = NULL,
  off_network_canalsditches = NULL,
  off_network_out = FALSE
)

Arguments

input_data

A list of input datasets created with nsink_prep_data.

off_network_lakes

Optional argument to set removal for waterbodies that are not part of the NHDPlus hydrologic network. Default value is to use the 75th percentile of removal from other lakes in the HUC. If another value is desired provide a single numeric ranging from 0 to 1.

off_network_streams

Optional argument to set removal for streams that are not part of the NHDPlus hydrologic network. Default value is to use the median removal from first order streams in the HUC. If another value is desired provide a single numeric ranging from 0 to 1.

off_network_canalsditches

Optional argument to set removal for canals and ditches that are not part of the NHDPlus hydrologic network. Default value is to use the 25th percentile of removal from third order streams in the HUC. If another value is desired provide a single numeric ranging from 0 to 1.

off_network_out

Optional argument to have the off network features returned on the output list.

Value

A list with three items, 1) a raster stack with one layer with the nitrogen removal, a second layer with the type of removal (e.g. hydric soils, lakes, streams), 2) a polygon representing removal from land, and 3) a polygon representing removal from the stream network, including stream removal, and lake removal. Optionally, the off network waterbodies may be returned if off_network_out == TRUE.

References

Kellogg, D. Q., Gold, A. J., Cox, S., Addy, K., & August, P. V. (2010). A geospatial approach for assessing denitrification sinks within lower-order catchments. Ecological Engineering, 36(11), 1596-1606. Link

Examples

if (FALSE) { # \dontrun{
library(nsink)
niantic_huc <- nsink_get_huc_id("Niantic River")$huc_12
niantic_data <- nsink_get_data(niantic_huc, data_dir = "nsink_data")
aea <- 5072
niantic_nsink_data <- nsink_prep_data(niantic_huc, projection = aea ,
                                      data_dir = "nsink_data")
removal <- nsink_calc_removal(niantic_nsink_data)
} # }