Finds 7-zip

This code is modified from https://github.com/jsta/nhdR/blob/master/R/utils.R and https://github.com/jsta/nhdR/blob/master/R/get.R to determine if 7 zip is available. If available it unzips a 7z zipfile to a destination directory. This avoids needing to use archive package which is only available via GitHub.

Usage

nsink_run_7z(zipfile, destdir, extract_again = FALSE)

Arguments

zipfile

The zipfile to be extracted

destdir

Where to put the extracted files

force

Whether or not to extract again if the destination files already exist

Author

Jemma Stachelek, jsta@lanl.gov